requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diffusers==0.30.3
edge-tts==6.1.12
feedparser==6.0.11
filelock==3.16.1
flake8==7.1.1
frozenlist==1.4.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gnewsclient==1.12
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mccabe==0.7.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
psutil==6.0.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pyflakes==3.2.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.3
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
sgmllib3k==1.0.0
SpeechRecognition==3.10.4
sympy==1.13.3
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.3
zipp==3.20.2
